Machine Learning Final Project - Loan Predictor
This is the final project for the Coursera IBM Certificate in Data Science course on Machine Learning.  The goal of this project is to use different machine learning algorithms known as K Nearest Neighbor, Decision Tree, Support Vector and Logistic Regression to predict whether a borrower will pay off or default on their loan and go into collections.  We then use Jaccard Score, F1 Score and Log Loss to rate the success of the predictions.